Documentation, Unit and Integration tests are added

+++ auction-service/src/main/java/com/db/auction/model/dto/crud/BidDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 public class BidDto extends BaseCrudDto<Bid, Long> {
     private long userUid;
-    private int auctionId;
+    private long auctionId;
     private int price;
 }
